requirements.txt
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.6
psycopg2==2.6.2
requests==2.11.1
